+++ app/src/main/java/com/cajetan/youtubeplus/utils/YouTubeData.java
@@ -29,6 +29,7 @@ import com.google.api.services.youtube.YouTubeScopes;
 import com.google.api.services.youtube.model.SearchListResponse;
 import com.google.api.services.youtube.model.SearchResult;
 import com.google.api.services.youtube.model.Video;
+import com.google.api.services.youtube.model.VideoListResponse;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -96,6 +97,12 @@ public class YouTubeData implements EasyPermissions.PermissionCallbacks {
         getResultsFromApi();
     }
 
+    public void receiveMostPopularResults() {
+        mRequestType = RequestType.MOST_POPULAR_REQUEST;
+
+        getResultsFromApi();
+    }
+
     private void getResultsFromApi() {
         if (!isGooglePlayServicesAvailable())
             acquireGooglePlayServices();
@@ -110,6 +117,8 @@ public class YouTubeData implements EasyPermissions.PermissionCallbacks {
                 new VideoSearchTask(mCredential).execute(searchQuery, searchPageToken);
             else if (mRequestType == RequestType.FAVOURITES_REQUEST)
                 new FavouritesTask(mCredential).execute(mVideoData.toArray(new VideoData[0]));
+            else if (mRequestType == RequestType.MOST_POPULAR_REQUEST)
+                new MostPopularTask(mCredential).execute();
         }
     }
 
@@ -333,6 +342,72 @@ public class YouTubeData implements EasyPermissions.PermissionCallbacks {
         }
     }
 
+    private class MostPopularTask extends AsyncTask<Void, Void, List<Video>> {
+        private com.google.api.services.youtube.YouTube mService;
+        private Exception mLastError = null;
+        private String mNextPageToken = null;
+        private String mPreviousPageToken = null;
+
+        MostPopularTask(GoogleAccountCredential credential) {
+            HttpTransport transport = AndroidHttp.newCompatibleTransport();
+            JsonFactory jsonFactory = JacksonFactory.getDefaultInstance();
+
+            mService = new com.google.api.services.youtube.YouTube.Builder(
+                    transport, jsonFactory, credential)
+                    .setApplicationName("YouTube Plus")
+                    .build();
+        }
+
+        @Override
+        protected List<Video> doInBackground(Void... args) {
+            Log.d(TAG, "Receiving search results..");
+
+            try {
+                 VideoListResponse response = mService.videos()
+                        .list("snippet,contentDetails")
+                        .setChart("mostPopular")
+                        .setRegionCode("US").execute();
+
+                this.mNextPageToken = response.getNextPageToken();
+                mPreviousPageToken = response.getPrevPageToken();
+
+                return response.getItems();
+            } catch (Exception e) {
+                Log.e("YouTubeData", "Exception: " + e.toString());
+                mLastError = e;
+                cancel(true);
+                return null;
+            }
+        }
+
+        @Override
+        protected void onPostExecute(List<Video> results) {
+            Log.d(TAG, "Passing the results..");
+
+            if (mActivity instanceof MostPopularListener)
+                ((MostPopularListener) mActivity).onMostPopularReceived(results, mNextPageToken, mPreviousPageToken);
+            else
+                throw new UnsupportedOperationException("Activity must implement FavouritesDataListener.");
+        }
+
+        @Override
+        protected void onCancelled() {
+            if (mLastError != null) {
+                if (mLastError instanceof GooglePlayServicesAvailabilityIOException) {
+                    Toast.makeText(mActivity, "GooglePlayServices not available.", Toast.LENGTH_LONG).show();
+                } else if (mLastError instanceof UserRecoverableAuthIOException) {
+                    mActivity.startActivityForResult(
+                            ((UserRecoverableAuthIOException) mLastError).getIntent(),
+                            YouTubeData.REQUEST_AUTHORIZATION);
+                } else {
+                    Log.e("YouTubeData", "Error occurred: " + mLastError.getMessage());
+                }
+            } else {
+                Log.e("YouTubeData", "Request cancelled");
+            }
+        }
+    }
+
     public void onParentActivityResult(int requestCode, int resultCode, Intent data) {
         switch (requestCode) {
             case REQUEST_GOOGLE_PLAY_SERVICES:
@@ -455,6 +530,7 @@ public class YouTubeData implements EasyPermissions.PermissionCallbacks {
         DATA_REQUEST,
         SEARCH_REQUEST,
         FAVOURITES_REQUEST,
+        MOST_POPULAR_REQUEST
     }
 
     public interface VideoDataListener {
@@ -468,5 +544,10 @@ public class YouTubeData implements EasyPermissions.PermissionCallbacks {
 
     public interface FavouritesDataListener {
         void onFavouritesReceived(List<Video> results);
+    }
+
+    public interface MostPopularListener {
+        void onMostPopularReceived(List<Video> results,
+                                   String nextPageToken, String previousPageToken);
     }
 }